requirements.txt
# Core UI
streamlit>=1.33


# Vision / ML
ultralytics>=8.2
opencv-python>=4.8
numpy>=1.24
Pillow>=10.0


# TensorFlow (choose one based on environment)
# tensorflow>=2.14
tensorflow-cpu>=2.14